+++ src/app/map/map.component.ts
@@ -1,8 +1,9 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, HostListener, OnInit } from '@angular/core';
 import { MapService } from '../shared/services/map.service';
 import * as L from 'leaflet';
 import { Config } from 'protractor';
 import { ConfigService } from '../shared/config/config.service';
+import { FormBuilder } from '@angular/forms';
 import { HttpClient, HttpHeaders } from '@angular/common/http';
 import { WorkflowService } from '../shared/services/workflow.service';
 import "leaflet/dist/images/marker-shadow.png";
@@ -19,11 +20,13 @@ import area from '@turf/area';
   templateUrl: './map.component.html',
   styleUrls: ['./map.component.scss']
 })
+
 export class MapComponent implements OnInit {
   public authHeader: HttpHeaders;
   private configSettings: Config;
   private messager: ToastrService;
   public clickPoint;
+  public basinCharacteristics;
   public currentZoom: number = 4;
   public latestDischarge: any;
 	public workflowLayers = [] as any;
@@ -32,7 +35,7 @@ export class MapComponent implements OnInit {
   public marker: L.Marker;
   public basin: any;
   public splitCatchmentLayer: any;
-  public traceLayer;
+  public traceLayerGroup;
   public fitBounds: L.LatLngBounds;
   public selectedWorkflow: Workflow;
   public loader: boolean;
@@ -41,14 +44,28 @@ export class MapComponent implements OnInit {
   public streamgageLayer: any;
   public streamgageLayerStatus: boolean;
   public workflowData: any;
+  public workflowForm: any;
   public count: number = 0;
+  public numFiresInClick: number = 0;
+  public firesinClick = [];
+  public selectedPerimeterHighlight;
   public foundFire:boolean = false; 
   public firePerimeterLayer;
-  public selectedPerimeters = [];
   public traceData = [];
+  public cursor = 'auto';
+  public selectedPerimeter: any;
+  public outputLayers: L.FeatureGroup;
+
+  @HostListener('document:click', ['$event']) 
+  clickout(event) 
+  { 
+    if (event.target.classList.contains("selectFire")) {
+      this.selectFire(event.path[2].innerHTML); 
+    }
+  }
 
   constructor(public _mapService: MapService, private _configService: ConfigService, private _http:
-    HttpClient, private _workflowService: WorkflowService, public toastr: ToastrService, private _loaderService: LoaderService, private _appService: AppService) { 
+    HttpClient, private _workflowService: WorkflowService, public toastr: ToastrService, private _loaderService: LoaderService, private _appService: AppService, private _fb: FormBuilder) { 
     this.configSettings = this._configService.getConfiguration();
     this.messager = toastr;
   }
@@ -63,6 +80,7 @@ export class MapComponent implements OnInit {
       renderer: L.canvas(),
       zoomControl: false
     });
+    this.outputLayers = new L.FeatureGroup().addTo(this._mapService.map);
 
     // Add basemap
     this._mapService.SetBaselayer(this._mapService.chosenBaseLayerName);
@@ -100,7 +118,7 @@ export class MapComponent implements OnInit {
     this._mapService.waterData.subscribe((wd: {}) => {
       this.latestDischarge = wd;
       if (this.latestDischarge) {
-        this.updatePopup(this.selectedSite, this.selectedPopup, this.selectedFeature);
+        this.updatePopup(this.selectedSite, this.selectedPopup, this.selectedFeature.properties['Name'], this.selectedFeature.properties.StationType.name, this.latestDischarge);
       }
     });
 
@@ -115,6 +133,11 @@ export class MapComponent implements OnInit {
     // Setting local click point variable
     this._mapService.clickPoint.subscribe((point: {}) => {
       this.clickPoint = point;
+    });
+
+    // Setting local basin characteristics variable
+    this._mapService.basinCharacteristics.subscribe((bc: {}) => {
+      this.basinCharacteristics = bc;
     });
 
     // On map click, set click point value
@@ -153,13 +176,20 @@ export class MapComponent implements OnInit {
     this._workflowService.selectedWorkflow.subscribe((res) => {
       this.selectedWorkflow = res;
       if (this.selectedWorkflow) {
-        if(this.selectedWorkflow.title != "Delineation") {
+        this.outputLayers = new L.FeatureGroup().addTo(this._mapService.map);
+        if (this.selectedWorkflow.title != "Delineation") {
           this.checkAvailableLayers();
         }
       }
-      if (!this.selectedWorkflow){
+      if (!this.selectedWorkflow) {
         this.removeWorkFlowLayers();
+        this.removeLayer(this.outputLayers);
       }
+    });
+
+    // Subscribe to the workflow form
+    this._workflowService.workflowForm.subscribe((workflowForm: any) => {
+      this.workflowForm = workflowForm;
     });
 
     // Subscribe to the form data
@@ -168,14 +198,51 @@ export class MapComponent implements OnInit {
       if (this.workflowData) {
         if (this.workflowData.title == "Delineation" || this.workflowData.title == "Fire Hydrology") {
           this.checkAvailableLayers();
+          // If Step #2 (Select Delineation Point) in the Delineation workflow has been completed
+          if (this.workflowData.title == "Delineation" && this.workflowData.steps[1].completed && !this.workflowData.steps[2].completed) {
+            // Check to see what Basin Characteristics are available for this point
+            await this.queryBasinCharacteristics();
+            // If at least one basin characteristic is available
+            // TODO: change these values from -9999.0 and -19.998 once this issue is resolved: https://code.usgs.gov/StreamStats/web-services-and-apis/cogQuery/lambdas/cq-lambda/-/issues/4
+            this.basinCharacteristics = this.basinCharacteristics.filter((basinCharacteristic) => basinCharacteristic.value != -9999.0 && basinCharacteristic.value != -19.998);
+            if (this.basinCharacteristics.length > 0) {
+              this._mapService.setBasinCharacteristics(this.basinCharacteristics);
+              let basinCharacteristicArray = [];
+              this.basinCharacteristics.forEach(basinCharacteristic=> {
+                basinCharacteristicArray.push(this._fb.group({
+                  text: basinCharacteristic.fcpg_parameter + ": " + basinCharacteristic.description,
+                  selected: false
+                }));
+              });
+              this.workflowForm.controls.steps.controls[2].controls.options.controls = basinCharacteristicArray;
+              this._workflowService.setWorkflowForm(this.workflowForm);
+            } else {
+              this.selectedWorkflow.steps[2].description = "No basin characteristics available at the clicked point."
+              this._mapService.setBasinCharacteristics(null);
+            }
+          }
+          if (this.workflowData.title == "Delineation" && this.workflowData.steps[2].completed) {
+            // If at least one basin characteristic was selected
+            if (this.workflowForm.controls.steps.controls[2].controls.options.controls.filter((checkboxBasinCharacteristic) => checkboxBasinCharacteristic.value.selected).length > 0) {
+              let selectedBasinCharacteristics = this.workflowForm.controls.steps.controls[2].controls.options.controls.filter(checkboxBasinCharacteristic => checkboxBasinCharacteristic.value.selected == true);
+              let selectedBasinCharacteristicCodes = selectedBasinCharacteristics.map(checkboxBasinCharacteristic => checkboxBasinCharacteristic.value.text.substr(0, checkboxBasinCharacteristic.value.text.indexOf(':')));
+              this.basinCharacteristics = this.basinCharacteristics.filter((basinCharacteristic) => selectedBasinCharacteristicCodes.includes(basinCharacteristic.fcpg_parameter));
+              this._mapService.setBasinCharacteristics(this.basinCharacteristics);
+
+            } else {
+              this._mapService.setBasinCharacteristics(null);
+            }
+          }
         }
         if (this.workflowData.title == "Fire Hydrology") {
-          if (this.workflowData.steps[1].name === "selectFireHydroBasin" && this.workflowData.steps[2].completed) {
-            this.queryBurnYear();
-          }
-          if (this.workflowData.steps[1].name === "selectFireHydroPerimeter"  && this.workflowData.steps[2].completed) {
-            this._loaderService.showFullPageLoad();
-            this.addTraceLayer(this.traceData);
+          if (this.workflowData.steps[1]) {
+            if (this.workflowData.steps[1].name === "selectFireHydroBasin" && this.workflowData.steps[2].completed) {
+              this.queryBurnYear();
+            }
+            if (this.workflowData.steps[1].name === "selectFireHydroPerimeter"  && this.workflowData.steps[2].completed) {
+              this._loaderService.showFullPageLoad();
+              this.addTraceLayer(this.traceData);
+            }
           }
         }
       }
@@ -187,6 +254,30 @@ export class MapComponent implements OnInit {
     // Subscribe to loader state
     this._loaderService.loaderState.subscribe((state: boolean) => {
       this.loader = state;
+    });
+
+    // Subscribe to current step
+    this._workflowService.currentStep.subscribe(step => {
+      if (step) {
+        this.cursor = step.cursor;
+      } else {
+        this.cursor = "auto";
+      }
+    });
+
+    // Get selected fire perimeters
+    this._mapService.selectedFirePerimeter.subscribe((perimeter) => {
+      this.selectedPerimeter = perimeter;
+      //remove old highlight
+      if (this.selectedPerimeterHighlight) {
+        var x_id = L.stamp(this.selectedPerimeterHighlight); // Retrieve the x layer ID
+        this.outputLayers.removeLayer(x_id);
+      }
+      // highlighted select perimeter 
+      this.selectedPerimeterHighlight = L.geoJSON(this.selectedPerimeter.Data);
+      this.selectedPerimeterHighlight.setStyle({color: "#ff0000"});
+      this.outputLayers.addLayer(this.selectedPerimeterHighlight);
+
     });
 
     // Load Layers
@@ -226,16 +317,16 @@ export class MapComponent implements OnInit {
             if (this.workflowData && this.workflowData.steps) {
               this.workflowData.steps[0].options.forEach((o: { text: string; selected: boolean; }) => {
                 if (o.selected == true) {
-                  if (o.text === "Query by Fire Perimeters" || o.text === "Query by Basin") {
+                  if (o.text === "Query by Fire Perimeter" || o.text === "Query by Basin") {
                     if (o.text === "Query by Basin") {
                       this.addLayers('NHD Flowlines', true);
                     }
                     this.addLayers('2000-2018 Wildland Fire Perimeters', true);
                     this.addLayers('2019 Wildland Fire Perimeters', true);
                     this.addLayers('2021 Wildland Fire Perimeters', true);
-                    this.addLayers('2022 Wildland Fire Perimeters', true);
+                    this.addLayers('Current Year Wildland Fire Perimeters', true);
                     this.addLayers('MTBS Fire Boundaries', true);
-                    this.addLayers('Burn Severity', true);
+                    this.addLayers('Burn Severity', false);
                   }
                 }
               })
@@ -347,15 +438,15 @@ export class MapComponent implements OnInit {
     }
   }
 
-  public updatePopup(site:any, popup:any, feature:any){
+  public updatePopup(site:any, popup:any, name:any, stationType:any, latestDischarge:any){
     //Set dynamic content for popup
     var SSgagepage = 'https://streamstatsags.cr.usgs.gov/gagepages/html/' + site + '.htm';
     var NWISpage = 'http://nwis.waterdata.usgs.gov/nwis/inventory/?site_no=' + site;
-    var innerHTML =  feature.properties['Name'] + ' ('  + site + ')' + '<hr><strong>Station Type</strong>: ' + 
-    feature.properties.StationType.name + '</br><strong>Discharge, cfs: </strong>' +
-    this.latestDischarge + '<br><strong>NWIS page: </strong><a href="' + 
-    NWISpage +' "target="_blank">link</a></br><strong>StreamStats Gage page: </strong><a href="' + 
-    SSgagepage + '" target="_blank">link</a></br>';
+    var innerHTML =  name + ' ('  + site + ')' + '<hr><strong>Station Type</strong>: ' + 
+    stationType + '<br>' +
+    (latestDischarge ? '<strong>Discharge, cfs: </strong>' + latestDischarge + '<br>' : '' ) +
+    '<strong>NWIS page: </strong><a href="' + 
+    NWISpage +' "target="_blank">link</a></br>';
     popup.setContent( innerHTML );
   }
 
@@ -390,7 +481,7 @@ export class MapComponent implements OnInit {
       if (this.basin) {  
         this.removeLayer(this.splitCatchmentLayer);  
         this.splitCatchmentLayer = L.geoJSON(this.basin.features[1]);
-        this.splitCatchmentLayer.addTo(this._mapService.map);
+        this.outputLayers.addLayer(this.splitCatchmentLayer);
         if (!this.splitCatchmentLayer.getBounds().isValid()) {
           this.createMessage("Error. Basin cannot be delineated.");
         } else {
@@ -408,7 +499,16 @@ export class MapComponent implements OnInit {
     this.removeLayer(this.marker);
     const content = '<div><b>Latitude:</b> ' + latlng.lat + '<br><b>Longitude:</b> ' + latlng.lng;
     this.marker = L.marker(latlng, {icon: RedIcon}).bindPopup(content).openPopup();
-    this._mapService.map?.addLayer(this.marker);
+    this.outputLayers.addLayer(this.marker);
+  }
+
+  public async queryBasinCharacteristics() {
+    this._loaderService.showFullPageLoad();
+    this.createMessage("Checking available basin characteristics. Please wait.");
+    let computedBasinCharacteristics = await this._mapService.queryPrecomputedBasinCharacteristics(this.clickPoint.lat, this.clickPoint.lng);
+    this._mapService.setBasinCharacteristics(computedBasinCharacteristics);
+    this.createMessage("Available basin characteristics were successfully checked.");
+    this._loaderService.hideFullPageLoad();
   }
 
   ////////////////////////////////
@@ -474,90 +574,196 @@ export class MapComponent implements OnInit {
   public onMouseClickFireHydroQueryFirePerimeter() { 
     this._loaderService.showFullPageLoad();
     this.count = 0;
+    this.numFiresInClick = 0;
+    this.selectedPerimeterHighlight = null;
     this.foundFire = false;
-    this.selectedPerimeters = [];
+    this.firesinClick = [];
+    this.createMessage("Querying fire perimeters in click point. Please wait.");
+
     Object.keys(this.workflowLayers).forEach(layerName => {
-      if (layerName === '2022 Wildland Fire Perimeters' || layerName === '2021 Wildland Fire Perimeters' || layerName === '2019 Wildland Fire Perimeters' || layerName === '2000-2018 Wildland Fire Perimeters') {
-        this.workflowLayers[layerName].query().nearby(this.clickPoint, 4).returnGeometry(true)
-          .run((error: any, results: any) => {
-            this.findFireFeatures(error,results,layerName);
+      if (this.activeWorkflowLayers.find(layer => layer.name === layerName)) {
+        if (this.activeWorkflowLayers.find(layer => layer.name === layerName).visible == true) {
+          if (layerName === 'Current Year Wildland Fire Perimeters' || layerName === '2021 Wildland Fire Perimeters' || layerName === '2019 Wildland Fire Perimeters' || layerName === '2000-2018 Wildland Fire Perimeters' || layerName === 'MTBS Fire Boundaries') {
+            this.workflowLayers[layerName].query().nearby(this.clickPoint, 4).returnGeometry(true)
+              .run((error: any, results: any) => {
+                this.findFireFeatures(error, results, layerName);
+              }
+            );
+         } else {
+            this.count ++;
+            this.checkCount(this.count, 6);
           }
-        );
-      } else if (layerName === 'MTBS Fire Boundaries') {
-        this.workflowLayers[layerName].identify().on(this._mapService.map).at(this.clickPoint).returnGeometry(true).tolerance(5)
-          .run(async (error: any, results: any) => {
-            this.findFireFeatures(error,results,layerName);
-          }
-        );
+        } else {
+          this.count ++;
+          this.checkCount(this.count, 6);
+        }
       }
     });
   }
 
-  public findFireFeatures(error,results,layerName) {
-    let popupcontent;
-    const shownFields = ['INCIDENTNAME', 'COMMENTS', 'GISACRES', 'FIRE_YEAR', 'CREATEDATE', 'ACRES', 'AGENCY', 'SOURCE', 'INCIDENT', 'FIRE_ID', 'FIRE_NAME', 'YEAR', 'STARTMONTH', 'STARTDAY', 'FIRE_TYPE', 
-    'POLY_INCIDENTNAME','POLY_GISACRES', 'POLY_DATECURRENT', 'IRWIN_FIRECAUSE', 'IRWIN_FIRECAUSEGENERAL', 'IRWIN_FIREDISCOVERYDATETIME','IRWIN_FIREOUTDATETIME','IRWIN_UNIQUEFIREIDENTIFIER'];
+  public findFireFeatures(error, results, layerName) {
+    
     if (error) {
       this.createMessage('Error occurred.','error');
       this._loaderService.hideFullPageLoad();
     } else if (results && results.features.length > 0) {
       this.foundFire = true;
-      results.features.forEach(feat => {
-        popupcontent = '<div class="popup-header"><b>' + layerName + ':</b></div><br>';
-        if (layerName === 'MTBS Fire Boundaries') {
-          let date = feat.properties.STARTMONTH + '/' + feat.properties.STARTDAY + '/' + feat.properties.YEAR;
-          if (date.indexOf('undefined') > -1) date = 'N/A';
+
+      if (results.features.length == 1) {
+        this.createContent(layerName, results.features[0]);
+        this.firesinClick.push({ 'Key': layerName, 'Data': results})
+      } else { // More than 1 of the same fire perimeter in click point
+        var temp = { 'crs': null, 'type': null, 'features': null }
+        var index = results.features.length;
+        for (let i = 0; i < index; i++) {         // need to split into separate fire perimeters
+          temp = { crs: results.crs, type: results.type, features: [results.features[i]] }
+          this.createContent(layerName, temp.features[0]);
+          this.firesinClick.push({ 'Key': layerName, 'Data': temp})
+          temp = { 'crs': null, 'type': null, 'features': null }
         }
-        Object.keys(feat.properties).forEach(prop => {
-          if (shownFields.indexOf(prop.toUpperCase()) > -1) {
-            let val = feat.properties[prop];
-            if (prop.toLowerCase().indexOf('date') > -1) {
-              val = new Date(val).toLocaleDateString();
-            }
-            popupcontent += '<b>' + prop.toUpperCase() + ':</b> ' + val + '<br>';
-          }
-        });
-        popupcontent += '<br>';
-        if (layerName === 'MTBS Fire Boundaries') {
-          this.firePerimeterLayer = L.geoJSON(feat.geometry);
-        } else if (layerName === '2022 Wildland Fire Perimeters' || layerName === '2021 Wildland Fire Perimeters' || layerName === '2019 Wildland Fire Perimeters' || layerName === '2000-2018 Wildland Fire Perimeters') {
-          const col = layerName.indexOf('Active') > -1 ? 'yellow' : 'red';
-          this.firePerimeterLayer = L.geoJSON(feat.geometry, {style: {color: col}});
-        }
-        this.firePerimeterLayer.addTo(this._mapService.map);
-        this.addBurnPoint(this.firePerimeterLayer.getBounds().getCenter(), popupcontent);
-      });
-      this.selectedPerimeters.push({ 'Key': layerName, 'Data': results})
-      this.traceData.push(results);
+      }
     } 
-    this._mapService.setSelectedPerimeters(this.selectedPerimeters);
     this.count ++;
-    this.checkCount(this.count, 5);
+    this.checkCount(this.count, 6);
+    if (this.firePerimeterLayer) {
+      this._mapService.map.fitBounds(this.firePerimeterLayer.getBounds(), {padding: [75,75]});
+    }
   }
 
-  public addTraceLayer(data) { 
-    var dataLength = (data.length);
-    data.forEach(async (trace, index) => {
-      var data = await this._mapService.trace(trace).toPromise();
-      if (data && data.features) {
-        data.features.forEach((feature) => {
-          if (feature.id == "flowlinesGeom") { // Only print out the flow lines
-            this.traceLayer = L.geoJSON(feature.geometry);
-          }
-        });   
+  public createContent(layerName, feat) {
+    let popupcontent;
+    var properties = {};
+    const shownFields = {
+      'INCIDENTNAME':"Incident Name",
+      'COMMENTS':"Comments",
+      'GISACRES':"Acres",
+      'FIRE_YEAR':"Fire Year",
+      'CREATEDATE':"Create Date",
+      'ACRES':"Acres",
+      'AGENCY':"Agency",
+      'SOURCE':"Source",
+      'INCIDENT':"Incident",
+      'FIRE_ID':"Fire ID",
+      'FIRE_NAME':"Fire Name",
+      'YEAR':"Year",
+      'STARTMONTH':"Start Month",
+      'STARTDAY':"Start Day",
+      'FIRE_TYPE':"Fire Type",					  
+      'POLY_INCIDENTNAME':"Incident Name",
+      'POLY_GISACRES':"Acres",
+      'POLY_DATECURRENT':"Modified Date",
+      'IRWIN_FIRECAUSE':"Fire Cause",
+      'IRWIN_FIRECAUSEGENERAL':"Fire Cause-General",
+      'IRWIN_FIREDISCOVERYDATETIME':"Fire Discovery Date Time",
+      'IRWIN_FIREOUTDATETIME':"Fire Out Date Time",
+      'IRWIN_UNIQUEFIREIDENTIFIER':"Unique Fire Identifier"
+    };
+
+    popupcontent = '<p hidden>' + this.numFiresInClick + '</p>'
+    popupcontent += '<div class="popup-header"><b>' + layerName + '</b></div><hr>';
+    if (layerName === 'MTBS Fire Boundaries') {
+      let date = feat.properties.STARTMONTH + '/' + feat.properties.STARTDAY + '/' + feat.properties.YEAR;
+      if (date.indexOf('undefined') > -1) date = 'N/A';
+    }
+    Object.keys(feat.properties).forEach(prop => {
+      if (Object.keys(shownFields).indexOf(prop.toUpperCase()) > -1) {
+        let label = shownFields[prop.toUpperCase()];
+        let val = feat.properties[prop];
+        if (prop.toLowerCase().indexOf('date') > -1) {
+          val = new Date(val).toLocaleDateString();
+        }
+        properties[label] = val;
+        popupcontent += '<b>' + label + ':</b> ' + val + '<br>';
       } 
-      this._mapService.setFirePerimetersLayers(this.firePerimeterLayer, this.traceLayer);
-      this.traceLayer.addTo(this._mapService.map);
-      this._mapService.map.fitBounds(this.traceLayer.getBounds(), { padding: [75,75] });
-      if (index + 1 == dataLength) {
-        this._loaderService.hideFullPageLoad();
+    });
+    popupcontent += '<br>'; 
+    popupcontent += '<center><button class="selectFire usa-button">Select Fire</button></center>'
+
+    this.firePerimeterLayer = L.geoJSON(feat.geometry);
+    this.addBurnPoint(this.firePerimeterLayer.getBounds().getCenter(), popupcontent);
+    feat.properties = properties;
+    this.numFiresInClick ++;
+  }
+
+
+  public selectFire(text) {
+    // if there was a fire perimeter previously selected, clear that from the map
+    let outputLayers = this.outputLayers;
+    this.outputLayers.eachLayer(function (layer) {
+      if (!(layer instanceof L.Marker)) {
+        outputLayers.removeLayer(layer);
       }
-    }); 
+    });
+    // figure out which perimeter they selected
+    var regex = /(?<=\>)(\d*)(?=\<\/p>)/g;
+    var result = text.match(regex)[0];
+    //  select new perimeter
+    this._mapService.setSelectedFirePerimeter(this.firesinClick[result]);
+    // set trace data 
+    this.traceData = this.firesinClick[result].Data;
+    // remove point markers for fire polygons
+    this.outputLayers.eachLayer(function (layer) {
+      if (layer instanceof L.Marker) {
+        outputLayers.removeLayer(layer);
+      }
+    });
+    // close all map popups
+    this._mapService.map.eachLayer(function (layer) {
+      layer.closePopup();
+    });
+  }
+
+  public async addTraceLayer(data) { 
+    var downstreamDist = this.workflowData.steps[2].options[0].text
+    this._mapService.setDownstreamDist(downstreamDist);
+
+      var response = await this._mapService.trace(data, downstreamDist).toPromise();
+
+      this.traceLayerGroup =  new L.FeatureGroup();
+      if (response) {
+        // show flowlines
+        if (response[1].features) {
+          response[1].features.forEach((feature) => {
+              this.traceLayerGroup.addLayer(L.geoJSON(feature.geometry));
+          });   
+        }
+        // show gages
+        if (response[2].features) {
+          let gageIcon;
+          this._mapService.setDownstreamGages(response[2].features); 
+          response[2].features.forEach((feature) => { 
+            if (feature.properties.active == true) {
+              gageIcon = L.divIcon({className: 'fireGageMarkerActive'});
+            } else if (feature.properties.active == false) {
+              gageIcon = L.divIcon({className: 'fireGageMarkerInactive'});
+            } else {
+              // TODO DELETE THIS LATER
+              gageIcon = L.divIcon({className: 'fireGageMarkerActive'});
+            }
+            let gageMarker = L.marker([feature.geometry.coordinates[1], feature.geometry.coordinates[0]], {icon: gageIcon});
+            let gageMarkerPopup = L.popup();
+            gageMarker.bindPopup(gageMarkerPopup);
+            if (feature.properties.Code) {
+              // Gage came from GageStatsServices
+              this.updatePopup(feature.properties.Code, gageMarkerPopup, feature.properties['Name'], feature.properties.StationType.name, null);
+            } else if (feature.properties.identifier) {
+              // Gage came from NavigationServices
+              this.updatePopup(feature.properties.identifier.substring(feature.properties.identifier.indexOf('-') + 1), gageMarkerPopup, feature.properties['name'], "Undefined", null);
+            }
+            this.traceLayerGroup.addLayer(gageMarker);
+          });
+        }
+      }
+      this._mapService.setFireTraceLayers(this.firePerimeterLayer, this.traceLayerGroup);
+      this.outputLayers.addLayer(this.traceLayerGroup);
+      this._mapService.map.fitBounds(this.traceLayerGroup.getBounds(), { padding: [75,75] });
+      this._loaderService.hideFullPageLoad();
   }
 
   public addBurnPoint(latlng, popupcontent) {
-    this.marker = L.marker(latlng).bindPopup(popupcontent).openPopup();
-    this._mapService.map?.addLayer(this.marker);
+    let firePerimeterIcon = L.divIcon({className: 'firePerimeterMarker'});
+    this.marker = L.marker(latlng, {icon: firePerimeterIcon}).bindPopup(popupcontent, {"autoClose": false}).openPopup();
+    this.outputLayers.addLayer(this.marker);
     this.marker.openPopup();
   }