+++ src/app/shared/interfaces/workflow/steps.ts
@@ -3,8 +3,10 @@ import { Options } from './options';
 export interface Steps {
     label: string;
     name: string;
+    description: string;
     value: string;
     type: string;
+    cursor: string; 
     validators?: any;
     options?: Array<Options>;
     //required?: boolean;